api/requirements.txt
fastapi
databases
aiomysql
sqlalchemy
uvicorn
bs4
selenium
mysql-connector-python
pyjwt
bcrypt
python-multipart